Refactor CBPS functions as classes and add an example +  package it (#1)
* Strucutre library and move some code to src

This commit adds a small template to build the library with code
in ./src. The cbps code is reduced to just a small class that can
be refactored to other applications or training mechanisms.

Some things to change are:
 - Figure out regularization. Added a not so fancy L1. Can we
   change the loss?
 - Apply other trainers.

* Test pytorch implementation using Lalonde

This commmit test the CBPS class with PyTorch with a simple example
using the Lalonde dataset.

Some things to work on:
 - regularization is in its infancy
 - numpy operations are buggy in weights calculation
 - implement some pytorch good practices when it comes to device.

* Add sys trick to run notebook in examples

Just add the sys.path trick to export libraries from the src/ folder
without moving the jupyter nobook to main.

* Add dimension reduction via SVD

We add some docs to the class, maybe something more technical would be
better? Also, when X is big, might be convenient to reduce the space and
factorize X to its minimal expression using a low-rank approach. I am
hacking things here as I am picking the singular-value threshold myself,
but there are basic alternatives (np.linalg.matrix_rank?), although
inspired by LoRA, I wanted to try something more "aggressive".

Bugs/TODO:
----
 - Maybe self.X can be defined via @property decorator instead
 - A logger might be a good idea

* Clean notebook