Add new standarized difference diagnostics (#2)
* Add new standarized difference diagnostics

Keeping everything in pytorch and do a standarized differences
diagnostics plot for the CBPS class. Also, change a bit the input
of the class to take a dataframe too.

Update notebook with the new stuff.

Bugs:
 - I haven't checked things thoroughly
 - I want to implement a ks stat.

* added third option to diagnose to plot a scatterplot of standardized vs unstandardized balance stats, and added an example with high dimensional data. Also added data dependencies.

* Change weights to return numpy array if wanted

Small changes to CBPS class to return numpy array if desired. Also, add
a new way to calculate variance when doing diagnostics.

---------

Co-authored-by: Apoorva Lal <lal.apoorva@gmail.com>